Move Click-related classes into plot module
These classes are used only in this module, so it makes sense to locate
them there.
